:recycle: ref: move notification context
